Correcting link syntax in Tutorial.pod
The link was missing a pipe (|) separator and was using a '/' in the link
text, which is disallowed by the POD formatting rules.  I've removed the
leading "t/" in the link text and hope that this is OK.  The sense of the
sentence is retained, the reader merely needs to know to look in the 't'
directory for the 900_run_features.t file.