start next "Scenario" even without empty lines
The parser discards empty lines after a multiline string.
So if the last step in a scenario has a multiline string,
then the next "Scenario" is not a valid step and causes a syntax error.
Self-referential example:

```
  Scenario: Step data in last step
    When I use step data in the last step
    Then I get an error:
      """
      -- Parse Error --

       Malformed step line
        at ...
        thrown by: [/.../Test/BDD/Cucumber/Parser.pm] line 277

      ...
      """

  Scenario: This line is considered a malformed step
```

This commit terminates the scenario steps when there is an empty line or
when a "Scenario:" or "Scenario Outline:" line is encountered.