Factor execution of scenario & background steps into its own function
By factoring the execution of steps into its own function and calling
that function for both the scenario and the background steps, the need
to recursively call into 'execute_scenario' has been eliminated.

Additionally, this allows the removal of conditionals which were
meant to be called *only* in case the call was /not/ the recursive
background call (but the toplevel 'scenario' call).

Now that execution is almost linear through the function, much less
brain power is required to understand the function's sequence of
events.