Add 'match' run mode to "pherkin"
This 'match' run mode serves to quickly identify which steps have
matching step functions and which ones don't.
The use-case for this functionality is that it doesn't require test
suites to set up any prerequisites (something which may consume time)
or to go through the exact steps of running the steps (which in case
of Selenium tests -- which I use T::B::C for) takes considerable time
too.