Extend the parser to allow multiple 'Examples:' per scenario
Gherkin allows multiple 'Examples:' sections per scenario; this
commit adds a Dataset model, extends the Scenario model and
changes the parser to generate these.

Before this commit, only an Examples section like this was supported

   Examples:
     | col1  |  col2 |
     | cv1   |  cv2  |

Note the lack of a "name" (text after the ':'), absent tags before
the 'Examples:' line and missing description block. None of those
were supported and all are added by this commit. The following
is now available:

  @day
  Examples: Day-examples
      These examples will be run during daytime
    | col1 | col2 |
    | dcv1 | dcv2 |

  @night
  Examples: Night-examples
      These examples will be run by night
    | col1 | col2 |
    | ncv1 | ncv2 |