Fit test description better in test output
The test output used to be:

```TAP
 not ok 18 - In Then the display should show 4567
 # Failed test 'In Then the display should show 4567'
 # at /opt/perl-5.30.1/lib/5.30.1/TAP/Parser/IteratorFactory.pm line 217.
```

which is confusing, both due to the quoting as well as the 'In Then' which
does not show immediately recognizable content to the user. This commit
changes it to

```TAP
 not ok 18 - Then the display should show 4567
 # Failed test 'Then the display should show 4567'
 # at /opt/perl-5.30.1/lib/5.30.1/TAP/Parser/IteratorFactory.pm line 217.
```

Which is actually much closer to what is happening, since the test
(which is itself a group of tests due to being a sub-test), fails.